small rework for update_database
